+++ pangea-sdk/v3/autogendoc/main.go
@@ -318,6 +318,5 @@ func gatherAstFieldList(fl []*ast.Field, fs *token.FileSet) []*AstField {
 func prettify(node interface{}, fs *token.FileSet) string {
 	var stringBuffer bytes.Buffer
 	printer.Fprint(&stringBuffer, fs, node)
-
-	return string(stringBuffer.Bytes())
+	return stringBuffer.String()
 }